backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
Werkzeug==3.0.3
python-dotenv==1.0.1
reportlab==4.2.2
gunicorn==22.0.0
flasgger==0.9.7.1

# CAD / DXF extraction
ezdxf==1.3.4

# OpenAI / ChatGPT integration
openai==2.12.0

# PostgreSQL driver
psycopg2-binary==2.9.10